refactor(parser): add mock generator and remove analyzer
Add MockGenerator
Remove ClassAnalyzer moving all the logics into ClassParser

BREAKING CHANGE: ClassAnalayzer has been removed and all the logic has been moved to ClassParser
