+++ packages/parser/src/lib/parser/class-parser.ts
@@ -1,54 +1,88 @@
-import { Inject, Service } from 'typedi';
-import { Faker, Class } from '@mockinbird/common';
-import { ParserConfig } from '../../types/types';
-import { ClassAnalyzer } from '../analyzer/class-analyzer';
-
-export interface ClassParser<TClass = any> {
-  parse(target: Class<TClass>): TClass;
-  parse(target: Class<TClass>, config?: ParserConfig<TClass>): TClass;
-  setFakerLocale(locale: Faker['locale']): void;
-}
+import { Container, Inject, Service } from 'typedi';
+import { ClassReflector, Property } from '@mockinbird/reflect';
+import { Class, Faker } from '@mockinbird/common';
+import { MutationsCallback, ParserConfig, ParsingStrategy } from '../types/types';
+import { ValueHandler } from '../types/value-handler.interface';
+import { EnumValueHandler } from '../handlers/enum-value-handler';
+import { ArrayValueHandler } from '../handlers/array-value-handler';
+import { SingleClassValueHandler } from '../handlers/single-class-value-handler';
+import { CallbackValueHandler } from '../handlers/callback-value-handler';
+import { ObjectLiteralValueHandler } from '../handlers/object-literal-value-handler';
+import { PrimitiveValueHandler } from '../handlers/primitive-value-handler';
+import { RegexValueHandler } from '../handlers/regex-value-handler';
 
 @Service()
 export class ClassParser<TClass = any> {
+  private readonly valueHandlers: Class<ValueHandler>[] = [
+    EnumValueHandler,
+    ArrayValueHandler,
+    SingleClassValueHandler,
+    CallbackValueHandler,
+    RegexValueHandler,
+    ObjectLiteralValueHandler,
+    PrimitiveValueHandler,
+  ];
+
   public constructor(@Inject('Faker') private readonly faker: Faker) {}
 
-  public setFakerLocale(locale: Faker['locale']): void {
-    this.faker.setLocale(locale);
+  private handlePropertyValue(property: Property): TClass | TClass[] {
+    for (const classHandler of this.valueHandlers) {
+      const handler = Container.get(classHandler);
+
+      if (handler.shouldHandle(property)) {
+        return handler.produceValue<TClass>(property);
+      }
+    }
   }
 
-  /**
-   * Return an object from the target class with all the properties
-   * decorated by the 'Mock' Decorator
-   *
-   * @param targetClass
-   */
-  public parse(targetClass: Class<TClass>): TClass;
-
-  /**
-   * Return an object from the target clxass with all the properties
-   * decorated by the 'Mock' Decorator
-   *
-   * Receive some extra configurations
-   *
-   * @param targetClass
-   * @param config
-   */
-  public parse(targetClass: Class<TClass>, config: ParserConfig<TClass>): TClass;
-
-  /**
-   * Return an object from the target class with all the properties
-   * decorated by the 'Mock' Decorator
-   *
-   * @param targetClass
-   * @param config
-   */
   public parse(targetClass: Class<TClass>, config: ParserConfig<TClass> = {}): TClass {
-    if (!targetClass) {
-      throw new Error(`Target class is 'undefined'`);
+    const classReflection = ClassReflector.getInstance().reflectClass(targetClass);
+
+    const { omit = [], pick = [] } = config;
+    let { mutations = {} } = config;
+    let strategy: ParsingStrategy;
+
+    if (omit.length) {
+      strategy = 'omit';
+    } else if (pick.length) {
+      strategy = 'pick';
     }
 
-    const analyzer = ClassAnalyzer.create<TClass>(targetClass);
-    return analyzer.analyzeProps(config);
+    if (omit.length && pick.length) {
+      throw new Error('Can not use pick and omit at the same time');
+    }
+
+    if (typeof mutations === 'function') {
+      mutations = (mutations as MutationsCallback<TClass>)(this.faker);
+    }
+
+    const deriveFromProps = (acc, property) => {
+      let value;
+
+      if (mutations.hasOwnProperty(property.name)) {
+        value = mutations[property.name];
+      }
+
+      if (strategy == 'pick') {
+        if (pick.includes(property.name)) {
+          return { ...acc, [property.name]: value || this.handlePropertyValue(property) };
+        }
+
+        return acc;
+      }
+
+      if (omit.includes(property.name) && strategy == 'omit') {
+        return acc;
+      }
+
+      return { ...acc, [property.name]: value || this.handlePropertyValue(property) };
+    };
+
+    const derivedProps = classReflection.reduce(deriveFromProps, {});
+    return Object.assign(new targetClass(), derivedProps);
+  }
+
+  public setLocale(locale: string): void {
+    this.faker.setLocale(locale);
   }
 }
